+++ java-io-base/src/test/java/nbbrd/io/text/ParserTest.java
@@ -257,6 +257,9 @@ public class ParserTest {
 
         assertThat(onURL().parse(null))
                 .isNull();
+
+        assertThat(onURL().parse("urlNotAbsolute"))
+                .isNull();
     }
 
     @Test